"conversion from unsigned int to const long requires a narrowing conversion" and why do you care so much to make this an error? GCC 9.3 is fine with it, it is taking an unsigned int (32-bit integer) and converting it to a signed long (32/64-bit signed integer). So what?

+++ src/gui/sdl_gui.cpp
@@ -638,8 +638,8 @@ namespace WLGUI {
 		void ObjSDLSurface::initFromSurface(void) {
 			viewport.w = abs(surface->w);
 			viewport.h = abs(surface->h);
-			scaleSrc = { viewport.w, viewport.h };
-			scaleDst = { viewport.w, viewport.h };
+			scaleSrc = { (long)viewport.w, (long)viewport.h };
+			scaleDst = { (long)viewport.w, (long)viewport.h };
 			ColorDescription.BitsPerPixel = surface->format->BitsPerPixel;
 			ColorDescription.BytesPerPixel = surface->format->BytesPerPixel;
